Moved the private classes into seperate files.
