+++ app/src/main/java/maustemies/halyttaja/AlarmManager.java
@@ -1,0 +1,20 @@
+package maustemies.halyttaja;
+
+/**
+ * Created by Maustemies on 17.12.2016.
+ */
+
+public class AlarmManager {
+
+    public boolean alarmOn = false;
+
+    public void StartAlarm() {
+        // TODO: Start the alarm
+        alarmOn = true;
+    }
+
+    public void StopAlarm() {
+        // TODO: Stop the alarm
+        alarmOn = false;
+    }
+}
